fix: address review comment 4460997262 - share internal field assertion
